requirements.txt
# Core framework
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-groq>=0.1.0

# Vector store
qdrant-client>=1.9.0,<1.12.0

# Embeddings and reranking
sentence-transformers>=2.7.0

# LLM
groq>=0.5.0

# UI
gradio>=4.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# For sparse vectors (BM25)
fastembed>=0.2.0

# Web search (Phase 2)
tavily-python>=0.3.0
